requirements.txt
fastapi
uvicorn[standard]

python-dotenv

mcp

langchain-community
langchain-openai
langsmith
langchain-mcp-adapters
langgraph

httpx
pytest
pytest-asyncio